pyproject.toml
# Build project configuration for release
[build-system]
requires = ["setuptools>=45", "setuptools_scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "aifn"
authors = [
    {name = "Weco AI Team", email = "dhruv@weco.ai"},
]
description = "Documentation for `aifn`, a client facing API for interacting with the WecoAI's AI functions."
readme = "README.md"
version = "0.2.0"
license = {text = "MIT"}
requires-python = ">=3.8"
dependencies = ["requests", "asyncio", "httpx[http2]", "pillow"]
keywords = ["AI", "LLM", "VLM", "AI functions", "Prompt Engineering", "NLP"]
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
    "License :: OSI Approved :: MIT License"
]

[project.urls]
Homepage = "https://github.com/WecoAI/aifn-python"

[project.optional-dependencies]
dev = ["ruff", "pytest-asyncio", "pytest-xdist", "build", "setuptools_scm"]
docs = ["jupyter", "black", "mkdocs", "mkdocs-material", "mkdocs-material-extensions", "mkdocstrings[crystal,python]", "mkdocs-minify-plugin"]

[tool.setuptools]
packages = ["aifn"]

[tool.setuptools_scm]

# Lint and format code
[tool.ruff]
fix = false
include = ["*.py"]
exclude = [".git"]
force-exclude = true
respect-gitignore = true
indent-width = 4
line-length = 127

[tool.ruff.lint]
ignore = ["E402"]

[tool.ruff.format]
docstring-code-format = true
indent-style = "space"
quote-style = "double"
skip-magic-trailing-comma = true
